requirements.txt
absl-py==2.1.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.7
Django==4.2.3
django-cors-headers==3.14.0
django-email-users==1.1.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
easy-colors==0.0.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.3.3
fsspec==2024.6.1
gast==0.6.0
gensim==4.3.1
google-pasta==0.2.0
grpcio==1.65.2
gunicorn==20.1.0
h5py==3.11.0
huggingface-hub==0.24.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
keras==3.4.1
langcodes==3.3.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
ml-dtypes==0.4.0
multidict==6.0.4
murmurhash==1.0.9
#mysqlclient==2.1.1
namex==0.0.8
nltk==3.8.1
numpy==1.24.3
openai==0.27.7
openapi-codec==1.3.2
opt-einsum==3.3.0
optree==0.12.1
packaging==23.1
pathy==0.10.1
preshed==3.0.8
protobuf==4.25.4
psutil==6.0.0
pydantic==1.10.9
Pygments==2.18.0
PyJWT==2.6.0
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0.1
regex==2024.7.24
requests==2.30.0
rich==13.7.1
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.10.1
send-mail==0.0.2
shortuuid==1.0.11
simplejson==3.19.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.17.0
thinc==8.1.10
threadpoolctl==3.1.0
tokenizers==0.19.1
tqdm==4.65.0
transformers==4.43.3
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vaderSentiment==3.3.2
wasabi==1.1.2
Werkzeug==3.0.3
whitenoise==6.5.0
wrapt==1.16.0
yarl==1.9.2
